make command for fix parsed pages

+++ app/Console/Commands/Book/FixPages.php
@@ -45,19 +45,21 @@ class FixPages extends Command
             $array = explode('||', $content);
             foreach ($array as $value){
                 $link = PageLink::query()->where('id', '=', $value)->first();
-                $page_num = explode('p=', $link->link);
-                $page = Page::query()
-                    ->where('book_id', '=', $link->book_id)
-                    ->where('page_number', '=', @end($page_num))
-                    ->first();
-                DB::transaction(function () use ($page, $link){
-                    if ($page){
-                        $page->delete();
-                    }
-                    $link->doParse = 1;
-                    $link->save();
-                });
-                echo $link->id.' - [FIXED]'."\n";
+                if ($link) {
+                    $page_num = explode('p=', $link->link);
+                    $page = Page::query()
+                        ->where('book_id', '=', $link->book_id)
+                        ->where('page_number', '=', @end($page_num))
+                        ->first();
+                    DB::transaction(function () use ($page, $link) {
+                        if ($page) {
+                            $page->delete();
+                        }
+                        $link->doParse = 1;
+                        $link->save();
+                    });
+                    echo $link->id . ' - [FIXED]' . "\n";
+                }
             }
         }
         echo '[COMPLETED]'."\n";
